refactor(tests): split test_api.py into per-resource test files
Extract shared fixtures (client, manufacturer, machine_model, etc.) and
SAMPLE_IMAGES into conftest.py, then split the monolithic test_api.py
(745 lines, 8 classes) into focused files by resource:

- test_api_models.py — TestModelsAPI
- test_api_titles.py — TestTitlesAPI, TestTitlesAllFacets
- test_api_manufacturers.py — TestManufacturersAPI
- test_api_people.py — TestPeopleAPI
- test_api_cache.py — TestAllEndpointCache
- test_api.py — TestSourcesAPI, TestSystemsAPI (remainder)
